requirements.txt
evdev
pyqt5 
hid 
python-uinput 
mss 
Pillow 
pyudev 
screeninfo
